split lexer tests into lexeme and program tests
Lexeme tests: check if the token type matches (e.g. integer, character etc.)
Program tests: check if a program is correctly splitted into tokens
