+++ syntax-antlr/src/test/java/lexer/features/TestEvents.java
@@ -1,5 +1,6 @@
-package lexer;
+package lexer.features;
 
+import lexer.LexerUtil;
 import org.antlr.v4.runtime.Token;
 import org.junit.Assert;
 import org.junit.Test;
