+++ syntax-antlr/src/test/java/lexer/lexemes/TestCharacter.java
@@ -1,5 +1,6 @@
-package lexer;
+package lexer.lexemes;
 
+import lexer.LexerUtil;
 import org.antlr.v4.runtime.Token;
 import org.junit.Assert;
 import org.junit.Test;
